Split the main example binary into examples
Separate fully featured demo from keybinding listing and event listener

Allows to make `gethostname` dev-dependency
