+++ Cargo.toml
@@ -11,10 +11,6 @@ repository = "https://github.com/nushell/reedline"
 [lib]
 doctest = true
 
-[[bin]]
-name = "reedline"
-path = "src/main.rs"
-
 [dependencies]
 chrono = "0.4.19"
 clipboard = { version = "0.5.0", optional = true }
@@ -30,7 +26,6 @@ strum_macros = "0.24"
 fd-lock = "3.0.3"
 rusqlite = { version = "0.28.0", optional = true }
 serde_json = { version = "1.0.79", optional = true }
-gethostname = { version = "0.2.3", optional = true }
 thiserror = "1.0.31"
 crossbeam = { version = "0.8.2", optional = true }
 
@@ -38,11 +33,13 @@ crossbeam = { version = "0.8.2", optional = true }
 tempfile = "3.3.0"
 pretty_assertions = "1.1.0"
 rstest = { version = "0.15.0", default-features = false }
+# For examples/demo.rs
+gethostname = "0.2.3"
 
 [features]
 system_clipboard = ["clipboard"]
 bashisms = []
 external_printer = ["crossbeam"]
-sqlite = ["rusqlite/bundled", "serde_json", "gethostname"]
-sqlite-dynlib = ["rusqlite", "serde_json", "gethostname"]
+sqlite = ["rusqlite/bundled", "serde_json"]
+sqlite-dynlib = ["rusqlite", "serde_json"]
